requirements.txt
Flask==3.0.0
Flask_Login==0.6.3
pycryptodome==3.19.0
PyExecJS==1.5.1
PyMySQL==1.1.0
